dashboards/requirements.txt
# Updated for Python 3.12
kafka-python==2.0.2
pyspark==3.5.1
prophet==1.1.5
xgboost==2.0.3
streamlit==1.32.2
psycopg2-binary==2.9.9
cassandra-driver==3.28.0
apache-airflow==2.8.0
apache-airflow-providers-http==4.5.0
apache-airflow-providers-postgres==5.7.1
#apache-airflow-providers-cassandra>=3.4.0
pendulum==2.1.2
requests==2.31.0
aiohttp==3.9.5
rich==13.7.1
tabulate==0.9.0
pyyaml==6.0.1
joblib==1.4.2
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
plotly